+++ drawio-custom-plugins/src/index.ts
@@ -2,6 +2,7 @@ import "./linkSelectedNodeWithData";
 import "./liveshare";
 import "./focus";
 import "./menu-entries";
+import "./notes";
 
 Draw.loadPlugin((ui) => {
 	(window as any).hediet_DbgUi = ui;